Move central index into context, selection state up
- Moved primer binding sites and sequence searching into utils
- removed unused CSS in index.css
- moved central index into context as well
- increase likelihood of jagged edges on overflowing annotations
- Moved selection state higher (into SeqViz) so it's again shared between both viewers
