+++ src/SeqViz/Circular/Selection/Selection.jsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 
-import { SelectionContext } from "../../handlers/selectionHandler";
+import { SelectionContext } from "../../handlers/selection";
 
 /**
  * renders the selection range of the plasmid viewer
