+++ src/SeqViz/Linear/InfiniteScroll/InfiniteScroll.jsx
@@ -1,5 +1,7 @@
-import { isEqual } from "lodash";
 import * as React from "react";
+import { isEqual } from "lodash";
+
+import CentralIndexContext from "../../handlers/centralIndex";
 
 /**
  * A wrapper around the seqBlocks. Renders only the seqBlocks that are
@@ -9,15 +11,17 @@ import * as React from "react";
  * seqBlocks should currently be shown
  */
 export default class InfiniteScroll extends React.PureComponent {
+  static contextType = CentralIndexContext;
+
   constructor(props) {
     super(props);
 
     this.state = {
-      visibleBlocks:
-        // start off with first 5 blocks shown
-        new Array(Math.min(5, props.seqBlocks.length))
-          .fill(null)
-          .map((_, i) => i)
+      // start off with first 5 blocks shown
+      visibleBlocks: new Array(Math.min(5, props.seqBlocks.length))
+        .fill(null)
+        .map((_, i) => i),
+      centralIndex: 0
     };
     this.scroller = React.createRef();
     this.insideDOM = React.createRef();
@@ -35,12 +39,9 @@ export default class InfiniteScroll extends React.PureComponent {
     }
 
     const { seqBlocks, size } = this.props;
-    const { visibleBlocks } = this.state;
+    const { centralIndex, visibleBlocks } = this.state;
 
-    if (
-      prevProps.seqBlocks[0].props.linearCentralIndex !==
-      seqBlocks[0].props.linearCentralIndex
-    ) {
+    if (this.context && centralIndex !== this.context.linear) {
       this.scrollToCentralIndex();
     } else if (
       !isEqual(prevProps.size, size) ||
@@ -81,7 +82,8 @@ export default class InfiniteScroll extends React.PureComponent {
   };
 
   /**
-   * Scroll to centralIndex
+   * Scroll to centralIndex. Likely from circular clicking on an element
+   * that should then be scrolled to in linear
    */
   scrollToCentralIndex = () => {
     const {
@@ -92,13 +94,13 @@ export default class InfiniteScroll extends React.PureComponent {
     } = this.props;
     const { visibleBlocks } = this.state;
     const { clientHeight, scrollHeight } = this.scroller.current;
+    const centralIndex = this.context.linear;
 
     // find the first block that contains the new central index
     const centerBlockIndex = seqBlocks.findIndex(
       block =>
-        block.props.firstBase <= block.props.linearCentralIndex &&
-        block.props.firstBase + block.props.bpsPerBlock >=
-          block.props.linearCentralIndex
+        block.props.firstBase <= centralIndex &&
+        block.props.firstBase + block.props.bpsPerBlock >= centralIndex
     );
 
     // build up the list of blocks that are visible just beneath this first block
@@ -132,7 +134,10 @@ export default class InfiniteScroll extends React.PureComponent {
     }
 
     if (!isEqual(newVisibleBlocks, visibleBlocks)) {
-      this.setState({ visibleBlocks: newVisibleBlocks });
+      this.setState({
+        visibleBlocks: newVisibleBlocks,
+        centralIndex: centralIndex
+      });
     }
   };
