+++ src/SeqViz/Linear/SeqBlock/Annotations/Annotations.jsx
@@ -177,7 +177,7 @@ class AnnotationRow extends React.PureComponent {
       linePath = `${topLeft} ${topRight}`;
     } else {
       let bottomRight = `L ${width} ${height}`; // flat right edge
-      if ((overflowRight && width > 10 * cW) || crossZero) {
+      if ((overflowRight && width > 2 * cW) || crossZero) {
         bottomRight = `
 				L ${width - cW} ${cH}
 				L ${width} ${2 * cH}
@@ -190,7 +190,7 @@ class AnnotationRow extends React.PureComponent {
       }
 
       let bottomLeft = `L 0 ${height} L 0 0`; // flat left edge
-      if (overflowLeft && width > 10 * cW) {
+      if (overflowLeft && width > 2 * cW) {
         bottomLeft = `
 				L 0 ${height}
 				L ${cW} ${3 * cH}
