+++ src/SeqViz/handlers/centralIndex.js
@@ -1,0 +1,13 @@
+import * as React from "react";
+
+const defaultCentralIndex = {
+  linear: 0,
+  circular: 0,
+  setCentralIndex: () => {}
+};
+
+/** Default central index context object */
+const CentralIndexContext = React.createContext(defaultCentralIndex);
+CentralIndexContext.displayName = "CentralIndexContext";
+
+export default CentralIndexContext;
